+++ app/src/main/java/io/ucoin/app/model/UcoinUd.java
@@ -1,5 +1,8 @@
 package io.ucoin.app.model;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 import io.ucoin.app.enumeration.DayOfWeek;
 import io.ucoin.app.enumeration.Month;
 
@@ -12,11 +15,11 @@ public interface UcoinUd extends SqlRow {
 
     Long time();
 
-    Long quantitativeAmount();
+    BigInteger quantitativeAmount();
 
-    Double relativeAmountThen();
+    BigDecimal relativeAmountThen();
 
-    Double relativeAmountNow();
+    BigDecimal relativeAmountNow();
 
     String currencyName();
